breakout bottom panel and modals
